Rewrite the functionality of commander
Commander is no longer meant to be write once and run anywhere. While this would be nice it is essentially a pipe dream and would cause a lot of overhead. Instead commands in DynamicGui in the future will likely just have their own compability layer or use CommanderSenders, since that could still be used. This was done due to too much boilerplate having to be written and then maintaining the code between multiple version would be an issue for just myself to maintain.