Refactor typing, use class type - 0.2.0
Dropped single method typing, using class typing instead. Had to enter generic hell for a bit but the result is you can now create a CommanderCommand<T> instead of the odd typing when creating a method with a specific type. This is a breaking change and as a result gets bumped to 0.2.0.